DCAC-153: Merge duplicate test config and utils classes.
