--- src/test/java/uk/gov/companieshouse/itemgroupstatusupdater/TestUtils.java
@@ -1,23 +1,0 @@
-package uk.gov.companieshouse.itemgroupstatusupdater;
-
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-
-public final class TestUtils {
-
-    public static final String MAIN_TOPIC = "echo";
-    public static final String RETRY_TOPIC = "echo-retry";
-    public static final String ERROR_TOPIC = "echo-error";
-    public static final String INVALID_TOPIC = "echo-invalid";
-
-    private TestUtils(){
-    }
-
-    public static int noOfRecordsForTopic(ConsumerRecords<?, ?> records, String topic) {
-        int count = 0;
-        for (ConsumerRecord<?, ?> ignored : records.records(topic)) {
-            count++;
-        }
-        return count;
-    }
-}
