+++ src/test/java/uk/gov/companieshouse/itemgroupstatusupdater/kafka/InvalidMessageRouterTest.java
@@ -5,7 +5,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.sameInstance;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.when;
-import static uk.gov.companieshouse.itemgroupstatusupdater.Constants.ITEM_GROUP_PROCESSED;
+import static uk.gov.companieshouse.itemgroupstatusupdater.util.TestConstants.ITEM_GROUP_PROCESSED;
 
 import java.util.Map;
 import org.apache.kafka.clients.producer.ProducerRecord;
